flaskr/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.10.3
amqp==5.1.1
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
bleach==6.0.0
blinker==1.6.2
cbor2==5.4.6
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post1
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
# Editable Git install with no remote (flaskr==1.0.0)
-e /home/lnv34/Downloads/Desktop/cookiecutter_django_project/flask/flaskdemoproject/flask-tutorial
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mkdocs-material-extensions==1.1.1
mongoengine==0.27.0
multidict==6.0.4
packaging==23.1
passlib==1.7.4
phonenumbers==8.13.9
phonenumberslite==8.13.10
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pypng==0.20220715.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
qrcode==7.4.2
requests==2.28.2
sendgrid==6.10.0
sentry-sdk==1.19.1
six==1.16.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.0
starkbank-ecdsa==2.2.0
twilio==8.0.0
typing-extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
vine==5.0.0
wcwidth==0.2.6
webauthn==1.8.0
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
WTForms-Components==0.10.5
yarl==1.8.2
zipp==3.15.0
